move RootGraph to topology
