+++ groalea/graphio.py
@@ -31,23 +31,11 @@ from .geometry import (TurtleState, FunctionalGeometry, rgb_color,
                        orientation, project3Dto2D, determinant, no_interior,
                        grotation, directionalTropism, orthogonalTropism, adjust_lu)
 
+from .topology import RootedGraph
+
 Vector3 = pgl.Vector3
 Vector4 = pgl.Vector4
 Color4Array = pgl.Color4Array
-
-
-class RootedGraph(PropertyGraph):
-    """ A general graph with a root vertex. """
-
-    def _set_root(self, root):
-        self._root = root
-
-    def _get_root(self):
-        return self._root
-
-    root = property(_get_root, _set_root)
-
-
 
 
 class Parser(object):
